backend/model-training-service/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
click==8.1.7
dnspython==2.7.0
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
idna==3.10
joblib==1.4.2
kafka-python==2.0.2
motor==3.6.0
numpy==2.2.0
pydantic==2.10.3
pydantic_core==2.27.1
pymongo==4.9.2
python-dotenv==1.0.1
scikit-learn==1.5.2
scipy==1.14.1
sniffio==1.3.1
starlette==0.41.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
uvicorn==0.32.1
